move QuantAlgorithm and QuantEncoding to constants.py
